int is fine

+++ ProjectEuler/ProjectEuler/Problem.h
@@ -7,4 +7,4 @@ using namespace std;
 int ProblemOne();
 long ProblemTwo();
 long long ProblemThree();
-long long ProblemFour();
+int ProblemFour();
